+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## v7.4.0 (planned on 01.09.2020)
-*Available in the `dev` branch*
+- arc: add set value by click feature
 
 ## v7.3.1 (planned on 18.08.2020)